[leetcode] - 350 intersect of two arrays (buggy)
This version is buggy for cases such as this:
    [1, 2] and [1, 1]
Depending on which one is selected as shorter and longer,
the answer will become either [1], or [1, 1].

We can fix this using weight. Build a binary tree as we sort
the array, and each number in the tree has a weight. Each
recurrence of the number means minus one from the number's weight.
If the weight becomes zero and yet we still find another recurrence,
this means we have exhaust all the recurrence of that number in
the other list, so the number cannot be added to the intersection.